chore: move app.js
